[🔨refactor] 회원가입에서 사용되는 useReducer 파일 분리
